docker: drop unblockUpdate
Automatically unblocking updates for a container only after it's started
is plain wrong. This as action, currently the only usage of blockUpdate,
doesn't necessarily mean the container is (re)started which means the
block is never removed. So remove the unblock list and track timestamps
until when a container should be blocked. This is possible as the events
as received by Docker container a timestamp of the event. So actions can
set a block until a certain point in time, and the start handler will
ignore all events which happened before this time.