docker: stop trying invoking actions on unknown container
Currently when a retry is started of invoking the actions this is tried
until the backoff stops. So this doesn't take into account that a / the
container can be stopped in the meantime as well which means it keeps
yielding errors. This fixes it by first checking if the container is
still known and if it isn't just returns succesfully and thus stops the
backoff.